fix(config/yaml): support yaml files which incompatible with json spec (#296)

+++ config/parser/yaml/yaml_test.go
@@ -27,7 +27,7 @@ func TestParser_Parse(t *testing.T) {
 					"name": "hello-kubernetes",
 				},
 				"spec": map[string]interface{}{
-					"replicas": 4,
+					"replicas": float64(4),
 				},
 			},
 		},
